requirements.txt
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
Django==5.2
django-ckeditor==6.7.2
django-js-asset==3.1.2
docutils==0.21.2
filetype==1.2.0
gunicorn==23.0.0
idna==3.10
packaging==24.2
Pygments==2.19.1
pypiwin32==223
pywin32==310
requests==2.32.3
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.4.0
